Disable Style/NumericPredicate
This cop recommends using `#positive?` over `> 0`, but the former is
only available in Ruby 2.3, which we practically don't use anywhere.

Enforcing comparisons is probably not a great idea since we do use
`#zero?` in some places (including `lib/aptible/tasks/rubocop.rb` in
this repo), so let's just disable it.